fix bug in good block marking. add tests.
